修改注册账号接口

+++ facehubSDKLib/src/main/java/com/azusasoft/facehubcloudsdk/api/FacehubApi.java
@@ -409,7 +409,7 @@ public class FacehubApi {
      */
     public void registerUser(String accessKey,
                              String sign,
-                             int deadLine,
+                             long deadLine,
                              final ResultHandlerInterface resultHandlerInterface) {
         String url = HOST + "/api/v1/users/";
         JSONObject params = new JSONObject();
